+++ src/gui/settings_dialog.py
@@ -52,8 +52,6 @@ class SparamviewerSettingsDialog(PygubuAppUI):
         self.impedance.trace_add('write', on_window_change)
 
         def on_editor_change(var, index, mode):
-            import logging
-            logging.error(self.ext_ed.get())
             Settings.ext_editor_cmd = self.ext_ed.get()
 
         self.ext_ed.trace_add('write', on_editor_change)
@@ -74,6 +72,12 @@ class SparamviewerSettingsDialog(PygubuAppUI):
         self.combobox_plotstyle['values'] = pyplot.style.available
         if Settings.plot_style in pyplot.style.available:
             self.combobox_plotstyle.current(pyplot.style.available.index(Settings.plot_style))
+        
+
+        def on_comment_expr_change(var, index, mode):
+            Settings.comment_existing_expr = self.comment_existing_expr.get()=='comment'
+        self.comment_existing_expr.set('comment' if Settings.comment_existing_expr else 'keep')
+        self.comment_existing_expr.trace_add('write', on_comment_expr_change)
         
         self.ext_ed.set(Settings.ext_editor_cmd)